+++ src/main/java/net/silentchaos512/gear/parts/PartSerializers.java
@@ -9,9 +9,11 @@ import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.parts.IGearPart;
 import net.silentchaos512.gear.api.parts.IPartSerializer;
 import net.silentchaos512.gear.api.parts.PartType;
+import net.silentchaos512.gear.config.Config;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.function.Supplier;
 
 public final class PartSerializers {
     private static final Map<ResourceLocation, IPartSerializer<?>> REGISTRY = new HashMap<>();
@@ -28,6 +30,7 @@ public final class PartSerializers {
         if (REGISTRY.containsKey(serializer.getName())) {
             throw new IllegalArgumentException("Duplicate gear part serializer " + serializer.getName());
         }
+        log(() -> "Registered serializer " + serializer.getName());
         REGISTRY.put(serializer.getName(), serializer);
         return serializer;
     }
@@ -36,6 +39,7 @@ public final class PartSerializers {
         String typeStr = JSONUtils.getString(json, "type");
         if (!typeStr.contains(":")) typeStr = SilentGear.RESOURCE_PREFIX + typeStr;
         ResourceLocation type = new ResourceLocation(typeStr);
+        log(() -> "deserialize " + id + " (type " + type + ")");
 
         IPartSerializer<?> serializer = REGISTRY.get(type);
         if (serializer == null) {
@@ -47,6 +51,7 @@ public final class PartSerializers {
     public static IGearPart read(PacketBuffer buffer) {
         ResourceLocation id = buffer.readResourceLocation();
         ResourceLocation type = buffer.readResourceLocation();
+        log(() -> "read " + id + " (type " + type + ")");
         IPartSerializer<?> serializer = REGISTRY.get(type);
         if (serializer == null) {
             throw new IllegalArgumentException("Unknown gear part serializer " + type);
@@ -54,10 +59,20 @@ public final class PartSerializers {
         return serializer.read(id, buffer);
     }
 
+    @SuppressWarnings("unchecked")
     public static <T extends IGearPart> void write(T part, PacketBuffer buffer) {
-        buffer.writeResourceLocation(part.getId());
-        buffer.writeResourceLocation(part.getSerializer().getName());
+        ResourceLocation id = part.getId();
+        ResourceLocation type = part.getSerializer().getName();
+        log(() -> "write " + id + " (type " + type + ")");
+        buffer.writeResourceLocation(id);
+        buffer.writeResourceLocation(type);
         IPartSerializer<T> serializer = (IPartSerializer<T>) part.getSerializer();
         serializer.write(buffer, part);
+    }
+
+    private static void log(Supplier<?> msg) {
+        if (Config.GENERAL.extraPartAndTraitLogging.get()) {
+            SilentGear.LOGGER.info(PartManager.MARKER, msg.get());
+        }
     }
 }