+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Config to enable additional logging for loading and syncing parts and traits
+### Fixed
+- Traits not loading on OS X
+
 ## [1.2.2] - 2019-07-17
 ### Added
 - New trait type, `silentgear:nbt_trait`. This will add arbitrary NBT to gear items when crafted.